Add continuous integration routines for codebase quality
