+++ .github/workflows/qlty.yml
@@ -1,0 +1,28 @@
+---
+name: Checking overall codebase quality
+
+on: [push, pull_request]
+
+jobs:
+  ci-qlty:
+    runs-on: ubuntu-latest
+    container: fedora:latest
+    strategy:
+      fail-fast: false
+    steps:
+      - name: Checkout the codebase current state
+        uses: actions/checkout@v4
+      - name: Install the system dependencies
+        uses: dnf install go
+      - name: Analyze codebase for potential issues and suspicious constructs
+        run: go vet ./...
+      - name: Ensure codebase adherence with standard Golang formatting style
+        run: go fmt ./...
+      - name: Install the base dependencies for staticcheck
+        run: go install honnef.co/go/tools/cmd/staticcheck@latest
+      - name: Evaluate codebase for deeper inefficiencies and potential bugs
+        run: ./$(go env GOPATH)/bin/staticcheck ./...
+      - name: Install the base dependencies for staticcheck
+        run: go install github.com/securego/gosec/v2/cmd/gosec@latest
+      - name: Inspect codebase for potential security vulnerabilities
+        run: ./$(go env GOPATH)/bin/gosec ./...
