+++ metasource/driver/util.go
@@ -1,11 +1,21 @@
 package driver
 
 import (
+	"crypto/rand"
 	"fmt"
 	"log/slog"
 	"os"
 	"strings"
 )
+
+func GenerateIdentity(length *int64) string {
+	var randBytes []byte
+
+	randBytes = make([]byte, *length/2)
+	_, _ = rand.Read(randBytes)
+
+	return fmt.Sprintf("%x", randBytes)
+}
 
 func InitPath(vers *string, loca *string) error {
 	var expt error
